requirements.txt
Django==5.1.3
djangorestframework==3.15.2
psycopg2==2.9.10
dj-database-url==2.3.0
python-dotenv==1.0.1
pytest-django==4.9.0
django-filter==24.3
gunicorn==23.0.0
drf-yasg==1.21.8
